refactor: update CompanyQueryService response structure and enhance response DTOs
