+++ company-service/src/main/java/com/winner/client/companyservice/company/application/service/CompanyQueryService.java
@@ -1,6 +1,7 @@
 package com.winner.client.companyservice.company.application.service;
 
 import com.winner.client.companyservice.company.presentation.dto.response.CompanyResponse;
+import com.winner.client.companyservice.company.presentation.dto.response.ListCompanyResponse;
 import java.util.List;
 import java.util.UUID;
 import org.springframework.transaction.annotation.Transactional;
@@ -8,12 +9,12 @@ import org.springframework.transaction.annotation.Transactional;
 @Transactional(readOnly = true)
 public interface CompanyQueryService {
 
-  List<CompanyResponse> getCompanyList();
+  ListCompanyResponse getCompanyList();
 
   CompanyResponse getCompany(UUID companyId);
 
   CompanyResponse getCompanyByIdAndCompanyName(UUID companyId, String companyName);
 
-  List<CompanyResponse> getCompanyListByCompanyName(String companyName);
+  ListCompanyResponse getCompanyListByCompanyName(String companyName);
 
 }
