+++ company-service/src/main/java/com/winner/client/companyservice/company/presentation/dto/response/ListCompanyResponse.java
@@ -3,9 +3,10 @@ package com.winner.client.companyservice.company.presentation.dto.response;
 import java.util.List;
 
 public record ListCompanyResponse(
-    List<CompanyResponse> companies
+    List<CompanyResponse> companies,
+    long totalCount
 ) {
-  public static ListCompanyResponse from(List<CompanyResponse> companies) {
-    return new ListCompanyResponse(companies);
+  public static ListCompanyResponse of(List<CompanyResponse> companies, Long totalCount) {
+    return new ListCompanyResponse(companies, totalCount);
   }
 }
